dy_ot/requirements.txt
opentelemetry-instrumentation
opentelemetry-exporter-otlp-proto-http
opentelemetry-sdk